executor: migrate TestTemporaryTableNoNetwork to testify (#27460)
