+++ executor/executor_test.go
@@ -8629,105 +8629,6 @@ func (s testSerialSuite) TestExprBlackListForEnum(c *C) {
 	c.Assert(checkFuncPushDown(rows, "index:idx(b, a)"), IsTrue)
 }
 
-func (s testSerialSuite) TestTemporaryTableNoNetwork(c *C) {
-	s.assertTemporaryTableNoNetwork(c, model.TempTableGlobal)
-	s.assertTemporaryTableNoNetwork(c, model.TempTableLocal)
-}
-
-func (s testSerialSuite) assertTemporaryTableNoNetwork(c *C, temporaryTableType model.TempTableType) {
-	var done sync.WaitGroup
-	defer done.Wait()
-
-	// Test that table reader/index reader/index lookup on the temporary table do not need to visit TiKV.
-	tk := testkit.NewTestKit(c, s.store)
-	tk1 := testkit.NewTestKit(c, s.store)
-
-	tk.MustExec("use test")
-	tk1.MustExec("use test")
-	tk.MustExec("drop table if exists normal, tmp_t")
-	tk.MustExec("create table normal (id int, a int, index(a))")
-	if temporaryTableType == model.TempTableGlobal {
-		tk.MustExec("set tidb_enable_global_temporary_table=true")
-		tk.MustExec("create global temporary table tmp_t (id int primary key, a int, b int, index(a)) on commit delete rows")
-	} else if temporaryTableType == model.TempTableLocal {
-		tk.MustExec("set tidb_enable_noop_functions=true")
-		tk.MustExec("create temporary table tmp_t (id int primary key, a int, b int, index(a))")
-	} else {
-		c.Fail()
-	}
-
-	c.Assert(failpoint.Enable("github.com/pingcap/tidb/store/mockstore/unistore/rpcServerBusy", "return(true)"), IsNil)
-	defer func() {
-		c.Assert(failpoint.Disable("github.com/pingcap/tidb/store/mockstore/unistore/rpcServerBusy"), IsNil)
-	}()
-
-	tk.MustExec("begin")
-	tk.MustExec("insert into tmp_t values (1, 1, 1)")
-	tk.MustExec("insert into tmp_t values (2, 2, 2)")
-
-	// Make sure the fail point works.
-	// With that failpoint, all requests to the TiKV is discard.
-	rs, err := tk1.Exec("select * from normal")
-	c.Assert(err, IsNil)
-	blocked := make(chan struct{}, 1)
-	ctx, cancelFunc := context.WithCancel(context.Background())
-	done.Add(1)
-	go func() {
-		defer done.Done()
-		_, _ = session.ResultSetToStringSlice(ctx, tk1.Se, rs)
-		blocked <- struct{}{}
-	}()
-	select {
-	case <-blocked:
-		c.Error("The query should block when the failpoint is enabled")
-	case <-time.After(200 * time.Millisecond):
-	}
-	cancelFunc()
-
-	// Check the temporary table do not send request to TiKV.
-	// PointGet
-	c.Assert(tk.HasPlan("select * from tmp_t where id=1", "Point_Get"), IsTrue)
-	tk.MustQuery("select * from tmp_t where id=1").Check(testkit.Rows("1 1 1"))
-	// BatchPointGet
-	c.Assert(tk.HasPlan("select * from tmp_t where id in (1, 2)", "Batch_Point_Get"), IsTrue)
-	tk.MustQuery("select * from tmp_t where id in (1, 2)").Check(testkit.Rows("1 1 1", "2 2 2"))
-	// Table reader
-	c.Assert(tk.HasPlan("select * from tmp_t", "TableReader"), IsTrue)
-	tk.MustQuery("select * from tmp_t").Check(testkit.Rows("1 1 1", "2 2 2"))
-	// Index reader
-	c.Assert(tk.HasPlan("select /*+ USE_INDEX(tmp_t, a) */ a from tmp_t", "IndexReader"), IsTrue)
-	tk.MustQuery("select /*+ USE_INDEX(tmp_t, a) */ a from tmp_t").Check(testkit.Rows("1", "2"))
-	// Index lookup
-	c.Assert(tk.HasPlan("select /*+ USE_INDEX(tmp_t, a) */ b from tmp_t where a = 1", "IndexLookUp"), IsTrue)
-	tk.MustQuery("select /*+ USE_INDEX(tmp_t, a) */ b from tmp_t where a = 1").Check(testkit.Rows("1"))
-	tk.MustExec("rollback")
-
-	// prepare some data for local temporary table, when for global temporary table, the below operations have no effect.
-	tk.MustExec("insert into tmp_t value(10, 10, 10)")
-	tk.MustExec("insert into tmp_t value(11, 11, 11)")
-
-	// Pessimistic lock
-	tk.MustExec("begin pessimistic")
-	tk.MustExec("insert into tmp_t values (3, 3, 3)")
-	tk.MustExec("insert ignore into tmp_t values (4, 4, 4)")
-	tk.MustExec("insert into tmp_t values (5, 5, 5) on duplicate key update a=100")
-	tk.MustExec("insert into tmp_t values (10, 10, 10) on duplicate key update a=100")
-	tk.MustExec("insert ignore into tmp_t values (10, 10, 10) on duplicate key update id=11")
-	tk.MustExec("replace into tmp_t values(6, 6, 6)")
-	tk.MustExec("replace into tmp_t values(11, 100, 100)")
-	tk.MustExec("update tmp_t set id = id + 1 where a = 1")
-	tk.MustExec("delete from tmp_t where a > 1")
-	tk.MustQuery("select count(*) from tmp_t where a >= 1 for update")
-	tk.MustExec("rollback")
-
-	// Check 'for update' will not write any lock too when table is unmodified
-	tk.MustExec("begin pessimistic")
-	tk.MustExec("select * from tmp_t where id=1 for update")
-	tk.MustExec("select * from tmp_t where id in (1, 2, 3) for update")
-	tk.MustExec("select * from tmp_t where id > 1 for update")
-	tk.MustExec("rollback")
-}
-
 func (s *testResourceTagSuite) TestResourceGroupTag(c *C) {
 	if israce.RaceEnabled {
 		c.Skip("unstable, skip it and fix it before 20210622")
