+++ testkit/testkit.go
@@ -19,6 +19,7 @@ package testkit
 import (
 	"context"
 	"fmt"
+	"strings"
 	"testing"
 
 	"github.com/pingcap/errors"
@@ -60,10 +61,10 @@ func (tk *TestKit) Session() session.Session {
 func (tk *TestKit) MustExec(sql string, args ...interface{}) {
 	res, err := tk.Exec(sql, args...)
 	comment := fmt.Sprintf("sql:%s, %v, error stack %v", sql, args, errors.ErrorStack(err))
-	tk.require.Nil(err, comment)
+	tk.require.NoError(err, comment)
 
 	if res != nil {
-		tk.require.Nil(res.Close())
+		tk.require.NoError(res.Close())
 	}
 }
 
@@ -99,6 +100,17 @@ func (tk *TestKit) ResultSetToResultWithCtx(ctx context.Context, rs sqlexec.Reco
 	rows, err := session.ResultSetToStringSlice(ctx, tk.session, rs)
 	tk.require.NoError(err, comment)
 	return &Result{rows: rows, comment: comment, assert: tk.assert, require: tk.require}
+}
+
+// HasPlan checks if the result execution plan contains specific plan.
+func (tk *TestKit) HasPlan(sql string, plan string, args ...interface{}) bool {
+	rs := tk.MustQuery("explain "+sql, args...)
+	for i := range rs.rows {
+		if strings.Contains(rs.rows[i][0], plan) {
+			return true
+		}
+	}
+	return false
 }
 
 // Exec executes a sql statement using the prepared stmt API
@@ -160,7 +172,7 @@ func (tk *TestKit) ExecToErr(sql string, args ...interface{}) error {
 
 func newSession(t *testing.T, store kv.Storage) session.Session {
 	se, err := session.CreateSession4Test(store)
-	require.Nil(t, err)
+	require.NoError(t, err)
 	se.SetConnectionID(testKitIDGenerator.Inc())
 	return se
 }
